Adds Node struct with new concepts of peerNodes.
