+++ node/http_req_res_utils.go
@@ -1,0 +1,42 @@
+package node
+
+import (
+	"net/http"
+	"encoding/json"
+	"io/ioutil"
+	"fmt"
+)
+
+func writeErrRes(w http.ResponseWriter, err error) {
+	jsonErrRes, _ := json.Marshal(ErrRes{err.Error()})
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(http.StatusInternalServerError)
+	w.Write(jsonErrRes)
+}
+
+func writeRes(w http.ResponseWriter, content interface{}) {
+	contentJson, err := json.Marshal(content)
+	if err != nil {
+		writeErrRes(w, err)
+		return
+	}
+
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(http.StatusOK)
+	w.Write(contentJson)
+}
+
+func readReq(r *http.Request, reqBody interface{}) error {
+	reqBodyJson, err := ioutil.ReadAll(r.Body)
+	if err != nil {
+		return fmt.Errorf("unable to read request body. %s", err.Error())
+	}
+	defer r.Body.Close()
+
+	err = json.Unmarshal(reqBodyJson, reqBody)
+	if err != nil {
+		return fmt.Errorf("unable to unmarshal request body. %s", err.Error())
+	}
+
+	return nil
+}
